+++ examples/basic-project/.browserslistrc
@@ -1,0 +1,1 @@
+chrome 55
