--- examples/basic-project/build.json
@@ -1,5 +1,0 @@
-{
-  "alias": {
-    "$$": "./.ice"
-  }
-}
