+++ examples/basic-project/package.json
@@ -18,6 +18,7 @@
     "@types/react": "^17.0.39",
     "@types/react-dom": "^17.0.11",
     "browserslist": "^4.19.3",
-    "regenerator-runtime": "^0.13.9"
+    "regenerator-runtime": "^0.13.9",
+    "webpack-bundle-analyzer": "^4.5.0"
   }
 }
