+++ packages/build-webpack-config/README.md
@@ -1,4 +1,4 @@
-# @builder/webpack-config
+# @ice/webpack-config
 
 This package providers basic webpack configuration.
 
@@ -6,7 +6,7 @@ This package providers basic webpack configuration.
 ## Usage
 
 ```js
-import { getWebpackConfig } from '@builder/webpack-config';
+import { getWebpackConfig } from '@ice/webpack-config';
 const config = { alias: {} };
 const rootDir = process.cwd();
 const webpackConfig = getWebpackConfig({ rootDir, config });
