+++ packages/ice/package.json
@@ -20,13 +20,13 @@
   "homepage": "https://next.ice.work",
   "dependencies": {
     "@builder/pack": "^0.6.0",
-    "@builder/webpack-config": "^3.0.0",
+    "@ice/webpack-config": "^1.0.0",
     "@ice/plugin-app": "^1.0.0",
     "@ice/plugin-auth": "^1.0.0",
     "@ice/route-manifest": "^1.0.0",
     "@ice/runtime": "^1.0.0",
     "address": "^1.1.2",
-    "build-scripts": "^2.0.0-7",
+    "build-scripts": "^2.0.0-14",
     "chalk": "^4.0.0",
     "commander": "^9.0.0",
     "consola": "^2.15.3",
