+++ tests/utils/build.ts
@@ -3,7 +3,7 @@ import process from 'process';
 import getPort from 'get-port';
 import Browser, { Page } from './browser';
 import getBuiltInPlugins from '../../packages/ice/src/getBuiltInPlugins';
-import createService from '../../packages/ice/src';
+import createService from '../../packages/ice/src/createService';
 import { fileURLToPath } from 'url';
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
