+++ tests/utils/start.ts
@@ -3,7 +3,7 @@ import getPort from 'get-port';
 import Browser, { Page } from './browser';
 import { Server } from 'http';
 import getBuiltInPlugins from '../../packages/ice/src/getBuiltInPlugins';
-import createService from '../../packages/ice/src';
+import createService from '../../packages/ice/src/createService';
 import { fileURLToPath } from 'url';
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
